Moved scraper parsing to own function
Cleans up the route a bit to make room for more routes (hint hint)
